+++ backend/localrep/models/organization.py
@@ -3,6 +3,7 @@ from django.db import models
 
 class ChannelOrganization(models.Model):
     organization_id = models.CharField(max_length=64)
+    address = models.CharField(max_length=200, default="")
     creation_date = models.DateTimeField()
     channel = models.CharField(max_length=100)