+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Backend hostname to the organization view (#1167)
+
 ## [0.21.0] 2022-07-05
 
 ### Added